+++ database/badgerdb/badger.go
@@ -20,6 +20,7 @@ func retry(dir string, originalOpts badger.Options) (*badger.DB, error) {
 
 func OpenDB(dir string, opts badger.Options) (*badger.DB, error) {
 	opts.Logger = nil
+	opts = opts.WithMetricsEnabled(true)
 	if db, err := badger.Open(opts); err != nil {
 		if strings.Contains(err.Error(), "LOCK") {
 			if db, err := retry(dir, opts); err == nil {