+++ database/dgraphdb/dgraph.go
@@ -3,6 +3,7 @@ package dgraphdb
 import (
 	"github.com/dgraph-io/dgo/v210"
 	"github.com/dgraph-io/dgo/v210/protos/api"
+	"github.com/spf13/viper"
 	"go.uber.org/zap"
 	"google.golang.org/grpc"
 )
@@ -17,7 +18,7 @@ type dgraph struct {
 func newClient() (*dgo.Dgraph, error) {
 	// Dial a gRPC connection. The address to dial to can be configured when
 	// setting up the dgraph cluster.
-	d, err := grpc.Dial("localhost:9080", grpc.WithInsecure())
+	d, err := grpc.Dial(viper.GetString("dgraph.host"), grpc.WithInsecure())
 	if err != nil {
 		return nil, err
 	}