+++ node/manager.go
@@ -1,4 +1,0 @@
-package node
-
-type Manager struct {
-}