+++ params/network.go
@@ -1,0 +1,6 @@
+package params
+
+const (
+	MainDiscoveryNode = "chain.rovergulf.net"
+	DefaultNodeAddr   = "127.0.0.1:9420"
+)